+++ modules/tensor_mechanics/test/tests/ad_anisotropic_creep/ad_aniso_creep_x_3d.i
@@ -1,14 +1,35 @@
 [Mesh]
-  type = GeneratedMesh
-  dim = 2
-  nx = 2
-  ny = 2
-  second_order = true
+  [gen]
+    type = GeneratedMeshGenerator
+    dim = 3
+    nx = 10
+    ny = 2
+    nz = 2
+    xmin = 0.0
+    ymin = 0.0
+    zmin = 0.0
+
+    xmax = 10.0
+    ymax = 1.0
+    zmax = 1.0
+  []
+  [corner_node]
+    type = ExtraNodesetGenerator
+    new_boundary = '100'
+    nodes = '3 69'
+    input = gen
+  []
+  [corner_node_2]
+    type = ExtraNodesetGenerator
+    new_boundary = '101'
+    nodes = '4 47'
+    input = corner_node
+  []
 []
 
 [GlobalParams]
-  displacements = 'disp_x disp_y'
-  volumetric_locking_correction = false
+  displacements = 'disp_x disp_y disp_z'
+  volumetric_locking_correction = true
 []
 
 [AuxVariables]
@@ -29,16 +50,7 @@
     family = MONOMIAL
   []
 []
-[Variables]
-  [disp_x]
-    order = SECOND
-    scaling = 1e-10
-  []
-  [disp_y]
-    order = SECOND
-    scaling = 1e-10
-  []
-[]
+
 [AuxKernels]
   [hydrostatic_stress]
     type = ADRankTwoScalarAux
@@ -48,22 +60,29 @@
   []
   [creep_strain_xx]
     type = ADRankTwoAux
-    rank_two_tensor = trial_creep_creep_strain
+    rank_two_tensor = creep_strain
     variable = creep_strain_xx
     index_i = 0
     index_j = 0
   []
   [creep_strain_xy]
     type = ADRankTwoAux
-    rank_two_tensor = trial_creep_creep_strain
+    rank_two_tensor = creep_strain
     variable = creep_strain_xy
     index_i = 0
     index_j = 1
   []
   [creep_strain_yy]
     type = ADRankTwoAux
-    rank_two_tensor = trial_creep_creep_strain
+    rank_two_tensor = creep_strain
     variable = creep_strain_yy
+    index_i = 1
+    index_j = 1
+  []
+  [sigma_xx]
+    type = ADRankTwoAux
+    rank_two_tensor = stress
+    variable = stress_xx
     index_i = 1
     index_j = 1
   []
@@ -72,70 +91,43 @@
 [Functions]
   [pull]
     type = PiecewiseLinear
-    x = '0 1e3 1e8'
-    y = '0 1e2 1e2'
+    x = '0 0.001 1.0'
+    y = '0 -4e1 -4e1'
   []
 []
 
 [Modules/TensorMechanics/Master]
   [all]
     strain = FINITE
-    add_variables = true
-    incremental = true
-    generate_output = 'elastic_strain_xx elastic_strain_yy elastic_strain_xy stress_xx stress_xy '
-                      'stress_yy'
+    generate_output = 'elastic_strain_xx stress_xx'
     use_automatic_differentiation = true
+    add_variables = true
   []
 []
 
 [Materials]
   [elasticity_tensor]
-    type = ADComputeElasticityTensor
-    fill_method = orthotropic
-    C_ijkl = '2.0e3 2.0e5 2.0e3 0.71428571e3 0.71428571e3 0.71428571e3 0.4 0.2 0.004 0.004 0.2 0.4'
+    type = ADComputeIsotropicElasticityTensor
+    youngs_modulus = 70000
+    poissons_ratio = 0.0
   []
 
   [elastic_strain]
     type = ADComputeMultipleInelasticStress
-    inelastic_models = "trial_creep trial_creep_two"
-    max_iterations = 500
-    absolute_tolerance = 1e-05
-  []
-
-  [trial_creep]
-    type = ADHillCreepStressUpdate
-    coefficient = 3e-18
-    n_exponent = 5
-    m_exponent = 0
-    activation_energy = 0
-    # F G H L M N
-    hill_constants = "0.5 0.5 0.3866 1.6413 1.6413 1.2731"
-    base_name = trial_creep
+    inelastic_models = "trial_creep_two"
+    max_iterations = 50
+    absolute_tolerance = 1e-16
   []
 
   [trial_creep_two]
     type = ADHillCreepStressUpdate
-    coefficient = 3e-18
-    n_exponent = 5
+    coefficient = 1e-16
+    n_exponent = 9
     m_exponent = 0
     activation_energy = 0
     # F G H L M N
-    hill_constants = "0.5 0.5 0.3866 1.6413 1.6413 1.2731"
-    base_name = trial_creep_two
-  []
-  [creep_one]
-    type = ADPowerLawCreepStressUpdate
-    coefficient = 1e-18
-    n_exponent = 5
-    activation_energy = 0
-    base_name = creep_one
-  []
-  [creep_nine]
-    type = ADPowerLawCreepStressUpdate
-    coefficient = 9e-24
-    n_exponent = 4
-    activation_energy = 0
-    base_name = creep_nine
+    hill_constants = "0.5 0.25 0.3866 1.6413 1.6413 1.2731"
+    max_inelastic_increment = 0.00003
   []
 []
 
@@ -143,50 +135,53 @@
   [no_disp_x]
     type = ADDirichletBC
     variable = disp_x
-    boundary = bottom
+    boundary = left
     value = 0.0
   []
 
   [no_disp_y]
     type = ADDirichletBC
     variable = disp_y
-    boundary = bottom
+    boundary = 100
+    value = 0.0
+  []
+
+  [no_disp_z]
+    type = ADDirichletBC
+    variable = disp_z
+    boundary = 101
     value = 0.0
   []
 
   [Pressure]
     [Side1]
-      boundary = top
+      boundary = right
       function = pull
     []
   []
-[]
 
-[Preconditioning]
-  [smp]
-    type = SMP
-    full = true
-  []
 []
 
 [Executioner]
   type = Transient
 
   solve_type = NEWTON
-  petsc_options_iname = '-ksp_gmres_restart -pc_type -sub_pc_type'
-  petsc_options_value = '101                asm      lu'
+  petsc_options_iname = '-pc_type -pc_factor_mat_solver_package'
+  petsc_options_value = 'lu     superlu_dist'
 
-  line_search = 'none'
-  nl_rel_tol = 1e-07
-  nl_abs_tol = 1.0e-15
+  nl_rel_tol = 1e-11
+  nl_abs_tol = 1.0e-14
   l_max_its = 90
-  num_steps = 40
-  dt = 5.0e1
+  num_steps = 25
+  dt = 1.0e-3
   start_time = 0
   automatic_scaling = true
 []
 
 [Postprocessors]
+  [matl_ts_min]
+    type = MaterialTimeStepPostprocessor
+  []
   [max_disp_x]
     type = ElementExtremeValue
     variable = disp_x
@@ -209,6 +204,24 @@
   [num_nonlin]
     type = NumNonlinearIterations
     outputs = console
+  []
+  [creep_strain_xx]
+    type = ElementalVariableValue
+    variable = creep_strain_xx
+    execute_on = 'TIMESTEP_END'
+    elementid = 39
+  []
+  [elastic_strain_xx]
+    type = ElementalVariableValue
+    variable = elastic_strain_xx
+    execute_on = 'TIMESTEP_END'
+    elementid = 39
+  []
+  [sigma_xx]
+    type = ElementalVariableValue
+    variable = stress_xx
+    execute_on = 'TIMESTEP_END'
+    elementid = 39
   []
 []
