+++ modules/tensor_mechanics/test/tests/ad_anisotropic_creep/gold/ad_aniso_creep_x_3d_out.csv
@@ -1,0 +1,27 @@
+time,creep_strain_xx,dt,elastic_strain_xx,matl_ts_min,max_disp_x,max_disp_y,max_hydro,sigma_xx
+0,0,0,0,0,0,0,0,0
+0.001,3.4351064258811e-06,0.001,0.00056799346498902,0.0087333538704554,0.0055951295382886,1.7190360352541e-15,13.333333333014,39.999999999043
+0.002,6.8702128522652e-06,0.001,0.0005679934649978,0.0087333538692049,0.0056287739568865,6.970639693351e-07,13.333333333231,39.999999999693
+0.003,1.0305319278637e-05,0.001,0.00056799346499748,0.0087333538691931,0.0056624184909688,1.3941267104198e-06,13.33333333322,39.99999999967
+0.004,1.3740425705002e-05,0.001,0.00056799346499739,0.00873335386918,0.005696063140628,2.0911882118167e-06,13.33333333322,39.999999999663
+0.005,1.7175532131364e-05,0.001,0.00056799346499734,0.0087333538691762,0.0057297079058606,2.7882484947843e-06,13.333333333222,39.999999999659
+0.006,2.0610638557729e-05,0.001,0.00056799346499728,0.0087333538692005,0.0057633527866647,3.4853075314959e-06,13.333333333219,39.999999999655
+0.007,2.4045744984093e-05,0.001,0.00056799346499739,0.0087333538691888,0.0057969977830444,4.1823653386879e-06,13.333333333219,39.999999999663
+0.008,2.7480851410458e-05,0.001,0.00056799346499733,0.0087333538691886,0.0058306428949985,4.8794219163974e-06,13.333333333219,39.999999999659
+0.009,3.0915957836826e-05,0.001,0.00056799346499746,0.0087333538691775,0.005864288122527,5.576477260612e-06,13.333333333218,39.999999999668
+0.01,3.4351064263194e-05,0.001,0.00056799346499743,0.0087333538691642,0.0058979334656313,6.2735313723161e-06,13.33333333322,39.999999999666
+0.011,3.7786170689561e-05,0.001,0.00056799346499739,0.0087333538691606,0.0059315789243103,6.9705842554873e-06,13.333333333219,39.999999999663
+0.012,4.1221277115934e-05,0.001,0.00056799346499738,0.0087333538691315,0.005965224498566,7.6676359037117e-06,13.333333333219,39.999999999663
+0.013,4.4656383542299e-05,0.001,0.00056799346499752,0.0087333538691386,0.0059988701883978,8.3646863217388e-06,13.33333333322,39.999999999672
+0.014,4.8091489968671e-05,0.001,0.00056799346499748,0.0087333538691433,0.0060325159938059,9.0617355077587e-06,13.333333333219,39.99999999967
+0.015,5.1526596395036e-05,0.001,0.00056799346499745,0.0087333538691324,0.0060661619147909,9.758783455635e-06,13.333333333217,39.999999999667
+0.016,5.4961702821399e-05,0.001,0.00056799346499736,0.0087333538691241,0.0060998079513543,1.0455830170405e-05,13.33333333322,39.999999999661
+0.017,5.8396809247763e-05,0.001,0.00056799346499733,0.008733353869149,0.0061334541034947,1.1152875654685e-05,13.333333333219,39.999999999659
+0.018,6.1831915674131e-05,0.001,0.00056799346499727,0.0087333538691485,0.0061671003712139,1.1849919914953e-05,13.333333333217,39.999999999655
+0.019,6.5267022100502e-05,0.001,0.0005679934649976,0.0087333538691509,0.0062007467545117,1.2546962933031e-05,13.33333333322,39.999999999678
+0.02,6.8702128526871e-05,0.001,0.0005679934649975,0.0087333538691412,0.0062343932533887,1.3244004727008e-05,13.333333333219,39.999999999671
+0.021,7.2137234953244e-05,0.001,0.00056799346499747,0.0087333538691441,0.0062680398678453,1.3941045284392e-05,13.33333333322,39.999999999669
+0.022,7.5572341379616e-05,0.001,0.00056799346499757,0.0087333538691386,0.0063016865978822,1.4638084611441e-05,13.33333333322,39.999999999676
+0.023,7.9007447805991e-05,0.001,0.00056799346499758,0.0087333538691265,0.0063353334434983,1.5335122704915e-05,13.333333333218,39.999999999677
+0.024,8.2442554232362e-05,0.001,0.00056799346499768,0.0087333538691187,0.0063689804046967,1.6032159567278e-05,13.33333333322,39.999999999684
+0.025,8.5877660658737e-05,0.001,0.0005679934649977,0.0087333538691081,0.0064026274814752,1.6729195196158e-05,13.33333333322,39.999999999685
