+++ modules/tensor_mechanics/test/tests/ad_anisotropic_creep/gold/ad_aniso_creep_y_3d_out.csv
@@ -1,0 +1,27 @@
+time,creep_strain_yy,dt,elastic_strain_yy,matl_ts_min,max_disp_x,max_disp_y,max_hydro,sigma_yy
+0,0,0,0,0,0,0,0,0
+0.001,1.5251574945778e-05,0.001,0.00055617699646912,0.0019670099715239,6.8793082356294e-17,0.0055951295382897,13.33333333301,39.999999999043
+0.002,3.0503149894834e-05,0.001,0.00055617699647959,0.0019670099711007,3.0948942299953e-06,0.0057445087245728,13.333333333334,40.000000000005
+0.003,4.5754724843892e-05,0.001,0.00055617699647959,0.0019670099710993,6.1897641717134e-06,0.0058938901890401,13.333333333333,40.000000000006
+0.004,6.1006299792965e-05,0.001,0.00055617699647954,0.0019670099710989,9.2846098231042e-06,0.0060432739318269,13.333333333333,40.000000000003
+0.005,7.6257874742012e-05,0.001,0.0005561769964795,0.0019670099710978,1.2379431181398e-05,0.0061926599529691,13.333333333333,39.999999999998
+0.006,9.150944969107e-05,0.001,0.00055617699647956,0.0019670099710986,1.5474228250001e-05,0.0063420482525007,13.333333333331,40.000000000003
+0.007,0.00010676102464014,0.001,0.00055617699647946,0.0019670099710944,1.8569001033667e-05,0.0064914388304567,13.333333333332,39.999999999997
+0.008,0.00012201259958918,0.001,0.00055617699647941,0.0019670099710926,2.1663749526139e-05,0.0066408316868716,13.333333333334,39.999999999992
+0.009,0.00013726417453825,0.001,0.00055617699647967,0.0019670099710966,2.4758473733192e-05,0.0067902268217794,13.333333333332,40.000000000011
+0.01,0.00015251574948734,0.001,0.00055617699647961,0.0019670099710908,2.785317364644e-05,0.0069396242352164,13.33333333333,40.000000000009
+0.011,0.00016776732443643,0.001,0.00055617699647958,0.0019670099710933,3.0947849272023e-05,0.0070890239272173,13.333333333333,40.000000000007
+0.012,0.00018301889938549,0.001,0.00055617699647953,0.0019670099710963,3.4042500615489e-05,0.0072384258978152,13.333333333333,40.000000000001
+0.013,0.00019827047433456,0.001,0.00055617699647968,0.0019670099710942,3.7137127667633e-05,0.0073878301470457,13.333333333332,40.000000000013
+0.014,0.00021352204928361,0.001,0.00055617699647952,0.0019670099710966,4.0231730434379e-05,0.0075372366749441,13.333333333333,40
+0.015,0.00022877362423266,0.001,0.00055617699647948,0.0019670099710952,4.3326308909696e-05,0.0076866454815446,13.333333333332,39.999999999997
+0.016,0.00024402519918169,0.001,0.00055617699647946,0.0019670099710908,4.6420863099752e-05,0.0078360565668815,13.333333333333,39.999999999994
+0.017,0.00025927677413072,0.001,0.00055617699647945,0.0019670099710931,4.9515393004716e-05,0.0079854699309908,13.333333333332,39.999999999994
+0.018,0.00027452834907975,0.001,0.00055617699647946,0.0019670099710937,5.2609898617241e-05,0.0081348855739067,13.333333333333,39.999999999994
+0.019,0.00028977992402877,0.001,0.00055617699647933,0.001967009971092,5.5704379945709e-05,0.008284303495664,13.333333333332,39.999999999984
+0.02,0.00030503149897783,0.001,0.00055617699647955,0.0019670099710965,5.8798836987596e-05,0.008433723696297,13.333333333332,40.000000000003
+0.021,0.00032028307392688,0.001,0.00055617699647947,0.0019670099710968,6.1893269742584e-05,0.008583146175841,13.333333333334,39.999999999997
+0.022,0.00033553464887591,0.001,0.00055617699647948,0.0019670099710943,6.4987678210002e-05,0.0087325709343301,13.333333333331,39.999999999995
+0.023,0.00035078622382498,0.001,0.00055617699647944,0.0019670099710954,6.808206239469e-05,0.0088819979717996,13.333333333332,39.999999999996
+0.024,0.00036603779877401,0.001,0.00055617699647946,0.0019670099710937,7.1176422293193e-05,0.0090314272882845,13.333333333332,39.999999999994
+0.025,0.00038128937372304,0.001,0.00055617699647943,0.0019670099710951,7.4270757904164e-05,0.0091808588838191,13.333333333332,39.999999999992
