+++ modules/tensor_mechanics/test/tests/ad_anisotropic_creep/gold/ad_aniso_creep_z_3d_out.csv
@@ -1,0 +1,27 @@
+time,creep_strain_zz,dt,elastic_strain_zz,matl_ts_min,max_disp_x,max_disp_y,max_hydro,sigma_zz
+0,0,0,0,0,0,0,0,0
+0.001,7.183161107578e-06,0.001,0.00056424541030732,0.0041764342398093,7.2985531030909e-17,4.585109838514e-15,13.333333333014,39.999999999043
+0.002,1.4366322214075e-05,0.001,0.00056424541029885,0.0041764342404381,1.0248475223739e-06,1.7368340028776e-06,13.333333332798,39.999999998375
+0.003,2.1549483320584e-05,0.001,0.00056424541029903,0.004176434240439,2.049692381403e-06,3.4736603569351e-06,13.333333332796,39.999999998388
+0.004,2.8732644427098e-05,0.001,0.00056424541029929,0.0041764342404291,3.0745345764362e-06,5.2104790629678e-06,13.333333332796,39.999999998406
+0.005,3.5915805533622e-05,0.001,0.00056424541029923,0.0041764342404204,4.0993741093777e-06,6.9472901184373e-06,13.333333332797,39.999999998403
+0.006,4.3098966640136e-05,0.001,0.00056424541029949,0.0041764342404159,5.1242109776484e-06,8.6840935169433e-06,13.333333332795,39.99999999842
+0.007,5.0282127746658e-05,0.001,0.00056424541029935,0.0041764342404062,6.1490451825669e-06,1.0420889275568e-05,13.333333332796,39.999999998411
+0.008,5.7465288853169e-05,0.001,0.00056424541029936,0.0041764342404043,7.1738767241848e-06,1.2157677375057e-05,13.333333332797,39.999999998411
+0.009,6.4648449959672e-05,0.001,0.00056424541029935,0.0041764342404032,8.1987056021443e-06,1.3894457832451e-05,13.333333332796,39.99999999841
+0.01,7.1831611066197e-05,0.001,0.00056424541029959,0.0041764342403929,9.2235318161794e-06,1.5631230635477e-05,13.333333332796,39.999999998428
+0.011,7.9014772172723e-05,0.001,0.00056424541029948,0.0041764342404017,1.0248355367372e-05,1.7367995794171e-05,13.333333332793,39.99999999842
+0.012,8.6197933279256e-05,0.001,0.00056424541029972,0.0041764342403976,1.1273176254632e-05,1.9104753298594e-05,13.333333332797,39.999999998437
+0.013,9.3381094385784e-05,0.001,0.00056424541029951,0.0041764342403884,1.2297994479167e-05,2.0841503157111e-05,13.333333332794,39.999999998423
+0.014,0.00010056425549231,0.001,0.00056424541029998,0.0041764342403711,1.3322810040823e-05,2.2578245367635e-05,13.333333332798,39.999999998456
+0.015,0.00010774741659883,0.001,0.00056424541029968,0.0041764342403778,1.4347622937487e-05,2.4314979920209e-05,13.333333332793,39.999999998434
+0.016,0.00011493057770535,0.001,0.00056424541029982,0.0041764342403786,1.5372433172307e-05,2.6051706828396e-05,13.333333332797,39.999999998444
+0.017,0.00012211373881188,0.001,0.00056424541029976,0.0041764342403804,1.6397240743054e-05,2.7788426086647e-05,13.333333332794,39.99999999844
+0.018,0.00012929689991841,0.001,0.00056424541029977,0.0041764342403812,1.7422045650868e-05,2.9525137695024e-05,13.333333332797,39.999999998441
+0.019,0.00013648006102493,0.001,0.00056424541029975,0.0041764342403749,1.844684789518e-05,3.1261841656311e-05,13.333333332795,39.999999998439
+0.02,0.00014366322213145,0.001,0.00056424541029964,0.0041764342403646,1.9471647476895e-05,3.2998537968871e-05,13.333333332796,39.999999998431
+0.021,0.00015084638323796,0.001,0.0005642454102999,0.0041764342403587,2.0496444393403e-05,3.4735226624738e-05,13.333333332794,39.999999998449
+0.022,0.00015802954434446,0.001,0.00056424541029966,0.0041764342403699,2.1521238647063e-05,3.6471907638151e-05,13.333333332798,39.999999998432
+0.023,0.00016521270545098,0.001,0.00056424541029981,0.0041764342403694,2.25460302379e-05,3.8208581002067e-05,13.333333332795,39.999999998443
+0.024,0.00017239586655749,0.001,0.00056424541029963,0.0041764342403689,2.3570819166347e-05,3.9945246716117e-05,13.333333332794,39.99999999843
+0.025,0.000179579027664,0.001,0.00056424541029994,0.0041764342403673,2.4595605429815e-05,4.1681904783059e-05,13.333333332795,39.999999998452
